Refactor image loading and encoding functions for improved error handling and clarity
- Updated `cargarImagen` to check for image file existence in a relative path.
- Enhanced error messages for better user feedback.
- Modified pixel encoding logic in `codificarImagen` to streamline pixel frequency counting.
- Improved file writing logic to ensure proper handling of binary data.
- Added checks for valid pixel counts during decoding in `decodificarArchivo`.
- Cleaned up output messages for consistency and clarity.
- Fixed minor formatting issues in user interface messages.

+++ ProyectoDelCurso/src/clases.cpp
@@ -112,7 +112,7 @@ bool Compare::operator()(HuffmanNode* a, HuffmanNode* b) {
 
 HuffmanTree::HuffmanTree(const vector<unsigned long>& freq) {
     priority_queue<HuffmanNode*, vector<HuffmanNode*>, Compare> pq;
-    for (int i = 0; i < freq.size(); i++) {
+    for (size_t i = 0; i < freq.size(); i++) {
         if (freq[i] > 0)
             pq.push(new HuffmanNode(i, freq[i]));
     }
